requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
Pillow==9.0.0
djangorestframework==3.13.1
django-ckeditor==6.2.0
